refactor: 카드 브랜드 관련 상수 및 함수 파일 분리
